Fixed #10: check that build-dir is not empty to consider resuse
Builder assumes that if the build-dir is present, it is on purpose and should not
clone pi-gen inside.
We were only testing for the existence of the folder which would be true when the
build-dir is not specified as it is inited to a tempdir.

We are now checking whether there are files inside as well.