Fixed #3: Add exfat data partition
export-image's prerun.sh has been rewritten to add an additional partition that is formatted to exfat.
It is mounted at `/data` and contains a `master_fs` flag file.

rootfs margin is set to 384Mi so it has a small space to use should it be needed ; since it won't be resized anymore.

A new `docker.fs` file is created on the third partition. It holds an empty ext4 filesytem that we'll be able to use for docker storage. Its size is configurable and default to 100MB. That docker fs is mounted at `/data/docker`

During build, exfat support is provided by exfat-fuse (only installed on build-system, not on target) and mkfs.exfat is provided (and installed) by exfatprogs (new exfat-related tools built by samsung and mounts via fuse). `exfat` module is loaded on start so we don't rely on fuse after build.

Similarily to how raspbian works, a first-time init script *resizes* the data partition. It actually recreates it as it can't be resized. This behavior is meant for base-image only as we expect base-image using creators to create full-size images.