Reduce disk writes
Reducing disk writes mostly through logs management:

We're instructing docker to send logs to journald so it's not storing files for it but `docker logs` can still be used.
Logs are also visible using `journalctl`

most of the logs should be going through systemd but regular syslog can still happen.
We are disabling the storage of syslog logs (so it's not affecting apps) by sending
all of them to the void.

main log store being systemd's journald, we are instructing it to use volatile storage.
Instead of storing on file in `/var/log/journal`, it stores it on RAM in `/run/log/journal`.
We are also requesting at most `30M` of journal storage.

Decided not to sync logs to disk (as does log2ram) as I don't think we'll ever need those
logs so it saves writes and complexity. Should one need the logs, it's still possible to
tweak the settings back.

Mounting the following as tmpfs:

- `/tmp`, `/var/tmp`
- `/var/cache`
- `/var/lib/misc`
- `/var/log`

Should prevent apps from writing logs to disk, from writting temp files to disk.

Decided not to set size limits on those mounts as this sounds difficult, fragile and error-prone.

As those are unexpected configurations on a system and because of the (hopefuly) rare
debug interventions, the motd has been changed and a logs-setup details file is added
so that admin knows where to find stuff and how to change setup should he need to.