Internet-connectivity checker
Knowing whether the system is *online* (in the sense of connected to public internet)
can be useful to many applications.
One of them (where it's mandatory) being the captive portal feature.
Instead of letting all those potential apps attempt to connect to internet to find out,
we can provide this service via a simple text file that can be bound to containers.

`/var/run/internet` is a text file which content is one of the following:

- **`online`**
- **`offline`**

Status is extrapolated from an UDP connection to Cloudflare Public DNS (`1.1.1.1`)
which is attempted every 30s via a systemd timer.

Beside this addition:

- it requests this to run before runtime-config so we are sure the file exists before creating containers