Added support for RTC hardware
Loads overlay for `i2c-rtc` device in order to support hardware clocks.
`hwclock` package is already installed and running on start so there is no
other configuration involved in benefiting from poweroff-resistent timekeeping.

It is users's responsibility to configure the HW clock.
Can be done with standard tools on any `--privileged` container.